// marketplace/src/types/custom-interface.ts
export interface ICountryDT {
  id: number
  name: string;
  properties: number;
  image: StaticImageData;
  tag: string;
  link: string;
}